app/requirements.txt
-e .
attrs==19.3.0
aws-cdk.aws-events==1.60.0
aws-cdk.aws-iam==1.60.0
aws-cdk.aws-kms==1.60.0
aws-cdk.aws-s3==1.60.0
aws-cdk.cloud-assembly-schema==1.60.0
aws-cdk.core==1.60.0
aws-cdk.cx-api==1.60.0
aws-cdk.region-info==1.60.0
cattrs==1.0.0
constructs==3.0.4
importlib-resources==3.0.0
jsii==1.11.0
pkg-resources==0.0.0
publication==0.0.3
python-dateutil==2.8.1
six==1.15.0
typing-extensions==3.7.4.3
zipp==3.1.0
